back-end v0.01

+++ Task 1/back-end/src/service/service.controller.ts
@@ -1,42 +1,82 @@
-import { Body, Controller, Delete, Get, Param, Post, Put } from '@nestjs/common';
+import { Body, Controller, Delete, Get, HttpStatus, Param, Post, Put, Req, Res } from '@nestjs/common';
 import { Service } from './schemas/service.schemas';
 import { ServiceService } from './service.service';
+import { Response, Request } from 'express';
+import { EventService } from 'src/event/event.service';
 
 @Controller('service')
 export class ServiceController {
-    constructor(private serviceService: ServiceService) { }
+    constructor(private serviceService: ServiceService, private eventService: EventService) { }
 
     @Get()
-    async getAllUService(): Promise<Service[]> {
-        return this.serviceService.findAll();
+    async getAllUService(
+        @Req() req: Request,
+    ): Promise<Service[]> {
+        if (req['user'].role === 'admin') {
+            return this.serviceService.findAll();
+        } else if (req['user'].role === 'user') {
+            const allServices: any = await this.serviceService.findAll();
+            const myEvent = await this.eventService.findAll(req['user']);
+            const oderService = [];
+            const newDate = new Date(Date.now());
+            allServices.forEach(service => {
+                if (!myEvent.some(myEvent => myEvent.serviceId == service._id)) {
+                    const date = new Date(service.endDate);
+                    if (newDate.getTime() < date.getTime()) {
+                        oderService.push(service); 
+                    }
+                }
+            });
+            return oderService;
+        }
     }
 
     @Post()
     async createService(
-        @Body()
-        service
-    ): Promise<Service> {
-        service.startDate = new Date(service.startDate);
-        service.endDate = new Date(service.endDate);
-        return this.serviceService.create(service)
+        @Req() req: Request,
+        @Res() res: Response,
+        @Body() service: Service,
+    ): Promise<Response> {
+        if (req['user'].role == 'admin') {
+            service.startDate = new Date(service.startDate);
+            service.endDate = new Date(service.endDate);
+            this.serviceService.create(service);
+            return res.status(HttpStatus.CREATED).json({ status: 'Ready' });
+        } else {
+            return res.status(HttpStatus.UNAUTHORIZED).json({ status: 'denied' });
+        }
     }
 
     @Put(':id')
     async updateService(
+        @Req() req: Request,
+        @Res() res: Response,
         @Param('id')
         id,
         @Body()
         service,
-    ): Promise<Service> {
-        service.startDate = new Date(service.startDate);
-        service.endDate = new Date(service.endDate);
-        return this.serviceService.updateById(id, service);
+    ): Promise<Response> {
+        if (req['user'].role == 'admin') {
+            service.startDate = new Date(service.startDate);
+            service.endDate = new Date(service.endDate);
+            this.serviceService.updateById(id, service);
+            return res.status(HttpStatus.OK).json({ status: 'Ready' });
+        } else {
+            return res.status(HttpStatus.UNAUTHORIZED).json({ status: 'denied' });
+        }
     }
     @Delete(':id')
     async deleteService(
+        @Req() req: Request,
+        @Res() res: Response,
         @Param('id')
         id,
-    ): Promise<Service> {
-        return this.serviceService.deleteById(id);
+    ): Promise<Response> {
+        if (req['user'].role == 'admin') {
+            this.serviceService.deleteById(id);
+            return res.status(HttpStatus.NO_CONTENT).json({ status: 'Ready' });
+        } else {
+            return res.status(HttpStatus.UNAUTHORIZED).json({ status: 'denied' });
+        }
     }
 }
